Fixing volatile and memory attributes for assembly.
For volatile:
You can prevent an asm instruction from being deleted by writing the
keyword volatile after the asm. [...] The volatile keyword indicates
that the instruction has important side-effects. GCC will not
delete a volatile asm if it is reachable.
and
An asm instruction without any output operands will be treated
identically to a volatile asm instruction.

And for memory:
This will cause GCC to not keep memory values cached in
registers across the assembler instruction and not optimize
stores or loads to that memory.

See also:
https://gcc.gnu.org/onlinedocs/gcc-4.7.2/gcc/Extended-Asm.html
http://stackoverflow.com/questions/14449141/the-difference-between-asm-asm-volatile-and-clobbering-memory