Add const_fn!{} macro to make functions const without duplication.
Many functions that can be const live in a generic context with bounds,
where `const fn` is not allowed without the "const_fn" feature.
In many places, functions duplicated becaue of this: A version with and
version without `const`.

This macro cleans that up by making it easy to define a function as
const only if possible, removing all duplication.

Also makes some more functions const using this macro.