Change signature of PhysToVirt::phys_to_virt
Instead of translating `PhysFrame`s to `*mut PageTable` pointers as before, it now performs a more general `PhysAddr` to `VirtAddr` conversion. Since the implementer must make sure that the returned address is valid and accessible, the trait is now unsafe to implement. Because of this requirement we also need to remove the generic implementation of this trait for closures.