fix: Enable use of OffsetPageTable with encryption bit in AMD SEV
AMD SEV use the C-Bit in the physical address to indicate, that the page
should point to encrypted memory in the virtual machine, so that it is
not readable by the host running the hypervisor.

This C-Bit is 1 << 51 for the Milan CPU architecture.

To construct an offset page table with the C-Bit set, the offset cannot
be a `VirtAddr`, because it would violate the constraints of a VirtAddr
for the higher bits.