Add SGX structures and methods for dynamic memory management
Introduce crate::structures::sgx for holding a flat list of SGX
micro-architectural structures, which are described in Intel SDM Vol. 3D,
chapter 34. Start with SECINFO, which describes page attributes for a
single page inside an SGX enclave (a protected memory region).

Add wrappers for EACCEPT, EACCEPTCOPY and EMODPE leaf functions of ENCLU
opcode, which are used to acknowledge, when a change is requested by the
host, and also modify enclave attributes. Add SGX return codes relevant
for the added functions.

Signed-off-by: Jarkko Sakkinen <jarkko@profian.com>