Make `GlobalDescriptorTable` const generic, but keep default size for `new`
Makes the `GlobalDescriptorTable` type const generic directly without introducing a new `BasicGlobalDescriptorTable` type.

To avoid breakage, we set a default size of 8 and change the `new` constructor function to always create a GDT of size 8. Creating emtpy GDTs with custom sizes is possible through a new `GlobalDescriptorTable<N>::empty` function.

This commit does _not_ duplicate the `from_raw_slice` function in the same way, so there might be some minor breakage for people that use this function. This was done intentionally so that users are prompted to decide for a GDT size that best matches the slice.