expose DEBUG_STR more directly
The constant is Sealed::DEBUG_STR was also already exposed to users
in its subtraits, but due to Sealed being unnamable reading DEBUG_STR
was somewhat difficult. Moving the associated constants from the
supertrait into the subtraits isn't a breaking change and makes it a
bit more convenient to reference.