implement functions for accessing CR8
CR8 aka the task priority register can be used to mask external
interrupts. If a non-zero value is written to the lower 4 bit of CR8,
all interrupts with an interrupt lower or equal to that number are
masked off. If CR8 contains 0, all external interrupts are enabled.

I've decided to implement this with an enum representing the 15
possible priority classes and used Option to represent the special 0
value. I've decided to make writing to CR8 safe because by itself, I
don't think CR8 can be used to cause UB. This is similar to our methods
for manipulating the IF flag.