Add page attribute table support
This adds support for the page attribute table MSR, and allows level 1
page table entries to have the `HUGE_PAGE` bit set. The `HUGE_PAGE` bit
is used as the `PAT` bit to index the table for 4KiB pages.

This commit is based on #529. Following this PR, it was brought up that
that PR was a breaking change because some users rely on
FrameError::HugeFrame. This PR is still a breaking change, but it tries
to make migrating easier for users that use this error variant.

I decided to go with Philipp's suggestion to add a is_level_1_entry
parameter. I'll concede that this doesn't feel particularly clean,
however callers usually already know the level of the page table entry
(at least the was the case for all the callers in this crate) and
having this function saves quite a lot of boilerplate (see #529 for how
calling code needs to be changed without this function detecting huge
pages). I tried looking through some projects on GitHub and it looks
like most of them don't have support for huge pages and just panic if
they encounter an error from this function. I think this is also
another good reason to keep the FrameError::HugeFrame variant around
because if a project ever starts to adopt huge pages, it will be very
obvious where code is failing opposed to frame() silently returning a
frame of the wrong size.

Co-Authored-By: Alison Davis <adavis628@proton.me>