add PageFaultErrorCode::HLAT
Some Intel CPUs have a feature called hypervisor-managed linear-address
translation (HLAT). The details of how HLAT paging works are a bit to
complicated for this commit message, but suffice it to say that with
HLAT, there are two root page tables, the regular PML4/PML5 and the
HLAT pointer (HLATP). If the page fault occured while translating with
the HLATP, the HLAT bit (7) is set in the error code. If the page fault
occured while translating with the regular PML4/PML5, the HLAT bit is
not set.